Fixed Merge issues
